made entire section page into a component, and added it to all relevent pages
