+++ client/admin-tool/src/components/PopOver.js
@@ -14,7 +14,7 @@ import Typography from '@mui/material/Typography';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 
-import "../css/SectionPopover.css";
+import "../css/Popover.css";
  
  /**
   *
