--- client/admin-tool/src/css/OrdinationIssue.css
@@ -1,3 +1,0 @@
-.orange-border {
-  border: 1px solid var(--darkorange);
-}
