+++ client/admin-tool/src/css/SectionWrapper.css
@@ -6,3 +6,7 @@
   flex-direction: row;
   justify-content: flex-start;
 }
+
+.orange-border {
+  border: 1px solid var(--darkorange);
+}
