+++ client/admin-tool/src/pages/About/EditSections.js
@@ -1,9 +1,23 @@
 import React from "react";
 
+import SectionWrapper from "../../components/Section/SectionWrapper";
+import {
+    fetchSectionsForPage,
+    addSection,
+    deleteSection,
+    updateSection,
+} from "../../util/requests/Sections";
+
 export default function EditSections() {
+
     return (
-        <div>
-            <p> This is the Edit Sections Page </p>
-        </div>
+        <SectionWrapper
+        PAGE="AboutUs"
+        pageTitle="About Us: Sections"
+        getItemsRequestCallback={fetchSectionsForPage}
+        addItemRequestCallback={addSection}
+        deleteItemRequestCallback={deleteSection}
+        updateItemRequestCallback={updateSection}
+        />
     );
 }
