+++ client/admin-tool/src/pages/Home/AddSection.js
@@ -1,9 +1,23 @@
 import React from "react";
 
+import SectionWrapper from "../../components/Section/SectionWrapper";
+import {
+    fetchSectionsForPage,
+    addSection,
+    deleteSection,
+    updateSection,
+} from "../../util/requests/Sections";
+
 export default function AddSection() {
+
     return (
-        <div>
-            <p> This is the Add Section Page </p>
-        </div>
+        <SectionWrapper
+        PAGE="Home"
+        pageTitle="Home: Sections"
+        getItemsRequestCallback={fetchSectionsForPage}
+        addItemRequestCallback={addSection}
+        deleteItemRequestCallback={deleteSection}
+        updateItemRequestCallback={updateSection}
+        />
     );
 }
