+++ client/admin-tool/src/pages/Resources/BuddhistCulture.js
@@ -1,20 +1,23 @@
-import React, { useState } from "react";
-import TextEditor from "../../components/TextEditor";
-import AddButton from "../../components/AddButton";
+import React from "react";
+
+import SectionWrapper from "../../components/Section/SectionWrapper";
+import {
+    fetchSectionsForPage,
+    addSection,
+    deleteSection,
+    updateSection,
+} from "../../util/requests/Sections";
 
 export default function BuddhistCulture() {
-    const [html, setHTML] = useState("");
 
     return (
-        <div style={{ marginLeft: "60px" }}>
-            <p> This is the BuddhistCulture Page </p>
-            <TextEditor editorUpdateCallback={setHTML} />
-            <p style={{ maxWidth: "60vw" }}>{html}</p>
-            <AddButton
-                onClickCallback={() => {
-                    alert("Click");
-                }}
-            />
-        </div>
+        <SectionWrapper
+        PAGE="BuddhistCulture"
+        pageTitle="Buddhist Culture"
+        getItemsRequestCallback={fetchSectionsForPage}
+        addItemRequestCallback={addSection}
+        deleteItemRequestCallback={deleteSection}
+        updateItemRequestCallback={updateSection}
+        />
     );
 }
